Updates from working on the problem, after class.
Without trying to be brief, I was able to make each of the items 1-4 happen
in 6 lines of code, each. Item 5 took 2 lines of code.

5 of 7 tasks, over 70% of the assignment can be done in 26 lines of code!

The brick wall will be more challenging. I'm going to start by making "bricks"
that are really just paddles that don't move. Sounds decent, no?